+++ Source/Engine/PointLight.h
@@ -11,8 +11,8 @@ public:
 
 	void Update(GLfloat deltaTime, int ID);
 
-	void UseLight(GLfloat ambientIntensityLocation, GLfloat ambientColorLocation, GLfloat diffuseInstensityLocation, 
-					GLfloat positionLocation, GLfloat attenuationVarsLocation);
+	void UseLight(GLuint ambientIntensityLocation, GLuint ambientColorLocation, GLuint diffuseInstensityLocation,
+		GLuint positionLocation, GLuint attenuationVarsLocation);
 
 	~PointLight();
