+++ Source/Engine/SpotLight.cpp
@@ -20,13 +20,13 @@ void SpotLight::Update(GLfloat deltaTime, int ID)
 {
 	Object::Update(deltaTime);
 
-	std::vector<GLfloat> uniforms;
+	std::vector<GLuint> uniforms;
 	shader->GetSpotLightLocationByID(uniforms, ID);
 
 	UseLight(uniforms[0], uniforms[1], uniforms[2], uniforms[3], uniforms[4], uniforms[5], uniforms[6]);
 }
 
-void SpotLight::UseLight(GLfloat ambientIntensityLocation, GLfloat ambientColorLocation, GLfloat diffuseInstensityLocation, GLfloat positionLocation, GLfloat directionLocation, GLfloat attenuationVarsLocation, GLfloat edgeLocation)
+void SpotLight::UseLight(GLuint ambientIntensityLocation, GLuint ambientColorLocation, GLuint diffuseInstensityLocation, GLuint positionLocation, GLuint directionLocation, GLuint attenuationVarsLocation, GLuint edgeLocation)
 {
 	glUniform3f(ambientColorLocation, color.x, color.y, color.z);
 	glUniform1f(ambientIntensityLocation, ambientIntensity);
